+++ src/themes/default/components/core/blocks/Footer/Footer.vue
@@ -55,17 +55,17 @@
                 {{ $t('About us') }}
               </h3>
               <div class="mt15">
-                <router-link class="cl-secondary" :to="localizedRoute('/i/about-us')" exact>
+                <router-link class="cl-secondary" :to="getLinkFor('/about-us')" exact>
                   {{ $t('About us (Magento CMS)') }}
                 </router-link>
               </div>
               <div class="mt15">
-                <router-link class="cl-secondary" :to="localizedRoute('/i/customer-service')" exact>
+                <router-link class="cl-secondary" :to="getLinkFor('/customer-service')" exact>
                   {{ $t('Customer service (Magento CMS)') }}
                 </router-link>
               </div>
               <div class="mt15">
-                <router-link class="cl-secondary" :to="localizedRoute('/store-locator')" exact>
+                <router-link class="cl-secondary" :to="getLinkFor('/store-locator')" exact>
                   {{ $t('Store locator') }}
                 </router-link>
               </div>
@@ -158,6 +158,7 @@
 </template>
 
 <script>
+import { currentStoreView, localizedRoute } from '@vue-storefront/core/lib/multistore'
 import CurrentPage from 'theme/mixins/currentPage'
 import LanguageSwitcher from '../../LanguageSwitcher.vue'
 import Newsletter from 'theme/components/core/blocks/Footer/Newsletter'
@@ -168,11 +169,20 @@ export default {
   mixins: [CurrentPage],
   name: 'MainFooter',
   computed: {
+    isStoreCodeEquals () {
+      return currentStoreView().storeCode === config.defaultStoreCode
+    },
     multistoreEnabled () {
       return config.storeViews.multistore
     },
     getVersionInfo () {
       return `v${process.env.__APPVERSION__} ${process.env.__BUILDTIME__}`
+    }
+  },
+  methods: {
+    getLinkFor (path) {
+      const route = this.isStoreCodeEquals ? `/i${path}` : path
+      return localizedRoute(route)
     }
   },
   components: {